+++ src/bpp/management/commands/usun_nieuzywane_charaktery_kbny.py
@@ -1,9 +1,8 @@
-# -*- encoding: utf-8 -*-
+import argparse
 
 from django.core.management import BaseCommand
-from django.db.models import CharField, TextField, Q
-from django.apps import apps
-from bpp.models import Sumy, Typ_KBN, Charakter_Formalny
+
+from bpp.models import Charakter_Formalny, Typ_KBN
 from bpp.util import usun_nieuzywany_typ_charakter
 
 
@@ -11,8 +10,8 @@ class Command(BaseCommand):
     help = "Usuwa nieuzywane charaktery formalne i typy KBN (stosuj po imporcie z DBF)"
 
     def add_arguments(self, parser):
-        parser.add_argument("--dry-run", action="store_true")
+        parser.add_argument("--dry-run", action=argparse.BooleanOptionalAction)
 
-    def handle(self, dry_run=False, *args, **options):
+    def handle(self, dry_run, *args, **options):
         usun_nieuzywany_typ_charakter(Typ_KBN, "typ_kbn", dry_run)
         usun_nieuzywany_typ_charakter(Charakter_Formalny, "charakter_formalny", dry_run)