+++ HISTORY.rst
@@ -4,6 +4,17 @@ Historia zmian
 
 .. towncrier release notes start
 
+Bpp 202310.1118 (2023-10-19)
+============================
+
+Usprawnienie
+------------
+
+- umożliwiaj importowanie punktów i dyscyplin źródeł z informacji z PBN,
+  umożliwiaj weryfikację źródeł po stronie PBN (ten sam ISSN, różne MNISWID,
+  brak informacji o dyscyplinach) (#1354)
+
+
 Bpp 202310.1116 (2023-10-01)
 ============================